Move SNS behaviour from Sinatras app.rb to a new class
app.rb was starting to inherit lots of SNS, and SES responsibilities
and knowledge.  Splitting this behaviour into a new class helps
keep app.rb generic to the routes it needs to know about.
